+++ Rdmp.UI/Menus/ExtractableCohortMenu.cs
@@ -4,7 +4,9 @@
 // RDMP is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
 // You should have received a copy of the GNU General Public License along with RDMP. If not, see <https://www.gnu.org/licenses/>.
 
+using System.Linq;
 using Rdmp.Core.DataExport.Data;
+using Rdmp.Core.Providers;
 using Rdmp.UI.CommandExecution.AtomicCommands;
 using Rdmp.UI.DataViewing;
 using Rdmp.UI.DataViewing.Collections;
@@ -23,6 +25,9 @@ namespace Rdmp.UI.Menus
             Items.Add("View TOP 100 identifiers",null, (s, e) => ViewTop100());
 
             Add(new ExecuteCommandDeprecate(args.ItemActivator, cohort, !cohort.IsDeprecated));
+
+            if (_activator.CoreChildProvider is DataExportChildProvider dx)
+                AddGoTo(dx.ExtractionConfigurations.Where(ec => ec.Cohort_ID == cohort.ID),"Extraction Configurations");
         }
         
         private void ViewTop100()