+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ...
 
+## [3.2.1-rc2] - 2019-10-18
+
+### Added 
+
+- Added GoTo from cohorts to Extraction Configuration(s)
+
+### Changed
+
+- View ThenVsNow Sql in right click context menu of data extractions is only evaluated when run (improves performance).  This results as the command always being enabled.
+
 ### Fixed
 
 - Fixed [bug in cross server query building](https://github.com/HicServices/RDMP/commit/a0c6223d1a7793bde4a67b368ae062e8bec3d960#diff-196fcda7990895e9f656c99602d1972b) (via cache) when joining patient index tables on one server to a main dataset on another
@@ -189,7 +199,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Attempting to deleting an Extractable Catalogue no longer shows an error and instead asks if you want to make it non extractable (then delete)
 - xmldoc are now shipped inside SourceCodeForSelfAwareness.zip (instead of side by side with the binary).  This avoids an issue where [Squirrel drops xmldoc files](https://github.com/Squirrel/Squirrel.Windows/issues/1323)
 
-
 ### Fixed 
 
 - Fixed bug in CLI (rdmp.exe) where yaml settings would override command line values for connection strings to platform databases
@@ -200,7 +209,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed Culture (e.g. en-us) not being passed correctly in DelimitedFlatFileAttacher
 - Fixed bug where Updater would show older versions of RDMP as installable 'updates'
 
-[Unreleased]: https://github.com/HicServices/RDMP/compare/3.2.1-rc1...develop
+[Unreleased]: https://github.com/HicServices/RDMP/compare/3.2.1-rc2...develop
+[3.2.1-rc2]: https://github.com/HicServices/RDMP/compare/3.2.1-rc1...3.2.1-rc2
 [3.2.1-rc1]: https://github.com/HicServices/RDMP/compare/3.2.0...3.2.1-rc1
 [3.2.0]: https://github.com/HicServices/RDMP/compare/v3.2.0-rc1...3.2.0
 [3.2.0-rc1]: https://github.com/HicServices/RDMP/compare/3.1.0...v3.2.0-rc1